requirements.txt
click==8.0.3
cycler==0.11.0
Cython==0.29.28
fonttools==4.29.1
imbalanced-learn==0.9.0
joblib==1.1.0
kiwisolver==1.3.2
liac-arff==2.5.0
matplotlib==3.5.1
nltk==3.7
np==1.0.2
numpy==1.18.5
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
regex==2022.1.18
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.62.3
